requirements.txt
flask
scikit-surprise
pandas
numpy<2.0
scikit-learn
